Enum zinc_hal_lpc11xx::ioregs::ADC_cr_start

pub enum ADC_cr_start {
    Stopped,
    Started,
    EdgePort0Pin2,
    EdgePortPin5,
    EdgeCT32B0_MAT0_1,
    EdgeCT32B0_MAT1_1,
    EdgeCT16B0_MAT0_1,
    EdgeCT16B0_MAT1_1,
}

Variants

Stopped

No start (this value should be used when clearing PDN to 0).

Started

Start conversion now.

EdgePort0Pin2

Start conversion when the edge selected by bit 27 occurs on PIO0_2/SSEL/CT16B0_CAP0.

EdgePortPin5

Start conversion when the edge selected by bit 27 occurs on PIO1_5/DIR/CT32B0_CAP0.

EdgeCT32B0_MAT0_1

Start conversion when the edge selected by bit 27 occurs on CT32B0_MAT0[1].

EdgeCT32B0_MAT1_1

Start conversion when the edge selected by bit 27 occurs on CT32B0_MAT1[1].

EdgeCT16B0_MAT0_1

Start conversion when the edge selected by bit 27 occurs on CT16B0_MAT0[1].

EdgeCT16B0_MAT1_1

Start conversion when the edge selected by bit 27 occurs on CT16B0_MAT1[1].