Enum zinc_hal_lpc11xx::ioregs::C_CAN_canstat_lec

pub enum C_CAN_canstat_lec {
    NoError,
    StuffError,
    FormError,
    AckError,
    Bit1Error,
    Bit0Error,
    CRCError,
    Unused,
}

Variants

NoError

No error.

StuffError

Stuff error: More than 5 equal bits in a sequence have occurred in a part of a received message where this is not allowed.

FormError

Form error: A fixed format part of a received frame has the wrong format.

AckError

AckError: The message this CAN core transmitted was not acknowledged.

Bit1Error

Bit1Error: During the transmission of a message (with the exception of the arbitration field), the device wanted to send a HIGH/recessive level (bit of logical value 1), but the monitored bus value was LOW/dominant.

Bit0Error

Bit0Error: During the transmission of a message (or acknowledge bit, or active error flag, or overload flag), the device wanted to send a LOW/dominant level (data or identifier bit logical value 0), but the monitored Bus value was HIGH/recessive. During busoff recovery this status is set each time a sequence of 11 HIGH/recessive bits has been monitored. This enables the CPU to monitor the proceeding of the busoff recovery sequence (indicating the bus is not stuck at LOW/dominant or continuously disturbed).

CRCError

CRCError: The CRC checksum was incorrect in the message received.

Unused

Unused: No CAN bus event was detected (written by the CPU).