Enum zinc_hal_lpc11xx::ioregs::I2C_mmctrl_match_all

pub enum I2C_mmctrl_match_all {
    MatchConfigured,
    MatchAnyAddress,
}

Variants

MatchConfigured

When this bit is cleared, an interrupt will only be generated when a match occurs to one of the (up-to) four address registers described above. That is, the module will respond as a normal slave as far as address-recognition is concerned.

MatchAnyAddress

When this bit is set to 1 and the I2C is in monitor mode, an interrupt will be generated on ANY address received. This will enable the part to monitor all traffic on the bus.