Struct zinc_hal_lpc11xx::ioregs::ADC

pub struct ADC;

10-bit ADC

Methods

impl ADC

fn cr() -> &'static mut ADC_cr

A/D Control Register. The ADCR register must be written to select the operating mode before A/D conversion can occur.

fn gdr() -> &'static mut ADC_gdr

A/D Global Data Register. Contains the result of the most recent A/D conversion.

fn stat() -> &'static mut ADC_stat

A/D Status Register. This register contains DONE and OVERRUN flags for all of the A/D channels, as well as the A/D interrupt flag.

fn inten() -> &'static mut ADC_inten

A/D Interrupt Enable Register. This register contains enable bits that allow the DONE flag of each A/D channel to be included or excluded from contributing to the generation of an A/D interrupt.

fn dr0() -> &'static mut ADC_dr0

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.

fn dr1() -> &'static mut ADC_dr1

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.

fn dr2() -> &'static mut ADC_dr2

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.

fn dr3() -> &'static mut ADC_dr3

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.

fn dr4() -> &'static mut ADC_dr4

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.

fn dr5() -> &'static mut ADC_dr5

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.

fn dr6() -> &'static mut ADC_dr6

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.

fn dr7() -> &'static mut ADC_dr7

A/D Channel n Data Register. This register contains the result of the most recent conversion completed on channel n.