Struct zinc_hal_lpc11xx::ioregs::ADC_cr

pub struct ADC_cr {
    // some fields omitted
}

Register CR: A/D Control Register. The ADCR register must be written to select the operating mode before A/D conversion can occur.

Methods

impl ADC_cr

fn get(&self) -> ADC_cr_Get

Fetch the value of the CR register.

fn sel(&self) -> u32

Get value of SEL field.

fn clkdiv(&self) -> u32

Get value of CLKDIV field.

fn burst(&self) -> ADC_cr_burst

Get value of BURST field.

fn clks(&self) -> ADC_cr_clks

Get value of CLKS field.

fn start(&self) -> ADC_cr_start

Get value of START field.

fn edge(&self) -> ADC_cr_edge

Get value of EDGE field.

fn ignoring_state(&self) -> ADC_cr_Update

Create new updater that ignores current value of the CR register.

fn set_sel<'a>(&'a mut self, new_value: u32) -> ADC_cr_Update<'a>

Set value of SEL field.

fn set_clkdiv<'a>(&'a mut self, new_value: u32) -> ADC_cr_Update<'a>

Set value of CLKDIV field.

fn set_burst<'a>(&'a mut self, new_value: ADC_cr_burst) -> ADC_cr_Update<'a>

Set value of BURST field.

fn set_clks<'a>(&'a mut self, new_value: ADC_cr_clks) -> ADC_cr_Update<'a>

Set value of CLKS field.

fn set_start<'a>(&'a mut self, new_value: ADC_cr_start) -> ADC_cr_Update<'a>

Set value of START field.

fn set_edge<'a>(&'a mut self, new_value: ADC_cr_edge) -> ADC_cr_Update<'a>

Set value of EDGE field.