Struct zinc_hal_lpc11xx::ioregs::ADC_stat_Get

pub struct ADC_stat_Get {
    // some fields omitted
}

STAT: A/D Status Register. This register contains DONE and OVERRUN flags for all of the A/D channels, as well as the A/D interrupt flag.

Methods

impl ADC_stat_Get

fn new(reg: &ADC_stat) -> ADC_stat_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn done(&self) -> u32

Get value of DONE field: These bits mirror the DONE status flags that appear in the result register for each A/D channel n.

fn overrun(&self) -> u32

Get value of OVERRUN field: These bits mirror the OVERRRUN status flags that appear in the result register for each A/D channel n. Reading ADSTAT allows checking the status of all A/D channels simultaneously.

fn adint(&self) -> bool

Get value of ADINT field: This bit is the A/D interrupt flag. It is one when any of the individual A/D channel Done flags is asserted and enabled to contribute to the A/D interrupt via the ADINTEN register.