Struct zinc_hal_lpc11xx::ioregs::CT16B0_ir

pub struct CT16B0_ir {
    // some fields omitted
}

Register IR: Interrupt Register (IR). The IR can be written to clear interrupts. The IR can be read to identify which of five possible interrupt sources are pending.

Methods

impl CT16B0_ir

fn get(&self) -> CT16B0_ir_Get

Fetch the value of the IR register.

fn mr0int(&self) -> bool

Get value of MR0INT field.

fn mr1int(&self) -> bool

Get value of MR1INT field.

fn mr2int(&self) -> bool

Get value of MR2INT field.

fn mr3int(&self) -> bool

Get value of MR3INT field.

fn cr0int(&self) -> bool

Get value of CR0INT field.

fn cr1int(&self) -> bool

Get value of CR1INT field.

fn ignoring_state(&self) -> CT16B0_ir_Update

Create new updater that ignores current value of the IR register.

fn set_mr0int<'a>(&'a mut self, new_value: bool) -> CT16B0_ir_Update<'a>

Set value of MR0INT field.

fn set_mr1int<'a>(&'a mut self, new_value: bool) -> CT16B0_ir_Update<'a>

Set value of MR1INT field.

fn set_mr2int<'a>(&'a mut self, new_value: bool) -> CT16B0_ir_Update<'a>

Set value of MR2INT field.

fn set_mr3int<'a>(&'a mut self, new_value: bool) -> CT16B0_ir_Update<'a>

Set value of MR3INT field.

fn set_cr0int<'a>(&'a mut self, new_value: bool) -> CT16B0_ir_Update<'a>

Set value of CR0INT field.

fn set_cr1int<'a>(&'a mut self, new_value: bool) -> CT16B0_ir_Update<'a>

Set value of CR1INT field.