Struct zinc_hal_lpc11xx::ioregs::CT16B0_pr

pub struct CT16B0_pr {
    // some fields omitted
}

Register PR: Prescale Register (PR). When the Prescale Counter (below) is equal to this value, the next clock increments the TC and clears the PC.

Methods

impl CT16B0_pr

fn get(&self) -> CT16B0_pr_Get

Fetch the value of the PR register.

fn pr(&self) -> u32

Get value of PR field.

fn ignoring_state(&self) -> CT16B0_pr_Update

Create new updater that ignores current value of the PR register.

fn set_pr<'a>(&'a mut self, new_value: u32) -> CT16B0_pr_Update<'a>

Set value of PR field.