Struct zinc_hal_lpc11xx::ioregs::CT32B0_emr

pub struct CT32B0_emr {
    // some fields omitted
}

Register EMR: External Match Register (EMR). The EMR controls the match function and the external match pins CT32B0_MAT[3:0].

Methods

impl CT32B0_emr

fn get(&self) -> CT32B0_emr_Get

Fetch the value of the EMR register.

fn em0(&self) -> bool

Get value of EM0 field.

fn em1(&self) -> bool

Get value of EM1 field.

fn em2(&self) -> bool

Get value of EM2 field.

fn em3(&self) -> bool

Get value of EM3 field.

fn emc0(&self) -> CT32B0_emr_emc0

Get value of EMC0 field.

fn emc1(&self) -> CT32B0_emr_emc1

Get value of EMC1 field.

fn emc2(&self) -> CT32B0_emr_emc2

Get value of EMC2 field.

fn emc3(&self) -> CT32B0_emr_emc3

Get value of EMC3 field.

fn ignoring_state(&self) -> CT32B0_emr_Update

Create new updater that ignores current value of the EMR register.

fn set_em0<'a>(&'a mut self, new_value: bool) -> CT32B0_emr_Update<'a>

Set value of EM0 field.

fn set_em1<'a>(&'a mut self, new_value: bool) -> CT32B0_emr_Update<'a>

Set value of EM1 field.

fn set_em2<'a>(&'a mut self, new_value: bool) -> CT32B0_emr_Update<'a>

Set value of EM2 field.

fn set_em3<'a>(&'a mut self, new_value: bool) -> CT32B0_emr_Update<'a>

Set value of EM3 field.

fn set_emc0<'a>(&'a mut self, new_value: CT32B0_emr_emc0) -> CT32B0_emr_Update<'a>

Set value of EMC0 field.

fn set_emc1<'a>(&'a mut self, new_value: CT32B0_emr_emc1) -> CT32B0_emr_Update<'a>

Set value of EMC1 field.

fn set_emc2<'a>(&'a mut self, new_value: CT32B0_emr_emc2) -> CT32B0_emr_Update<'a>

Set value of EMC2 field.

fn set_emc3<'a>(&'a mut self, new_value: CT32B0_emr_emc3) -> CT32B0_emr_Update<'a>

Set value of EMC3 field.