Struct zinc_hal_lpc11xx::ioregs::CT32B0_mcr

pub struct CT32B0_mcr {
    // some fields omitted
}

Register MCR: Match Control Register (MCR). The MCR is used to control if an interrupt is generated and if the TC is reset when a Match occurs.

Methods

impl CT32B0_mcr

fn get(&self) -> CT32B0_mcr_Get

Fetch the value of the MCR register.

fn mr0i(&self) -> CT32B0_mcr_mr0i

Get value of MR0I field.

fn mr0r(&self) -> CT32B0_mcr_mr0r

Get value of MR0R field.

fn mr0s(&self) -> CT32B0_mcr_mr0s

Get value of MR0S field.

fn mr1i(&self) -> CT32B0_mcr_mr1i

Get value of MR1I field.

fn mr1r(&self) -> CT32B0_mcr_mr1r

Get value of MR1R field.

fn mr1s(&self) -> CT32B0_mcr_mr1s

Get value of MR1S field.

fn mr2i(&self) -> CT32B0_mcr_mr2i

Get value of MR2I field.

fn mr2r(&self) -> CT32B0_mcr_mr2r

Get value of MR2R field.

fn mr2s(&self) -> CT32B0_mcr_mr2s

Get value of MR2S field.

fn mr3i(&self) -> CT32B0_mcr_mr3i

Get value of MR3I field.

fn mr3r(&self) -> CT32B0_mcr_mr3r

Get value of MR3R field.

fn mr3s(&self) -> CT32B0_mcr_mr3s

Get value of MR3S field.

fn ignoring_state(&self) -> CT32B0_mcr_Update

Create new updater that ignores current value of the MCR register.

fn set_mr0i<'a>(&'a mut self, new_value: CT32B0_mcr_mr0i) -> CT32B0_mcr_Update<'a>

Set value of MR0I field.

fn set_mr0r<'a>(&'a mut self, new_value: CT32B0_mcr_mr0r) -> CT32B0_mcr_Update<'a>

Set value of MR0R field.

fn set_mr0s<'a>(&'a mut self, new_value: CT32B0_mcr_mr0s) -> CT32B0_mcr_Update<'a>

Set value of MR0S field.

fn set_mr1i<'a>(&'a mut self, new_value: CT32B0_mcr_mr1i) -> CT32B0_mcr_Update<'a>

Set value of MR1I field.

fn set_mr1r<'a>(&'a mut self, new_value: CT32B0_mcr_mr1r) -> CT32B0_mcr_Update<'a>

Set value of MR1R field.

fn set_mr1s<'a>(&'a mut self, new_value: CT32B0_mcr_mr1s) -> CT32B0_mcr_Update<'a>

Set value of MR1S field.

fn set_mr2i<'a>(&'a mut self, new_value: CT32B0_mcr_mr2i) -> CT32B0_mcr_Update<'a>

Set value of MR2I field.

fn set_mr2r<'a>(&'a mut self, new_value: CT32B0_mcr_mr2r) -> CT32B0_mcr_Update<'a>

Set value of MR2R field.

fn set_mr2s<'a>(&'a mut self, new_value: CT32B0_mcr_mr2s) -> CT32B0_mcr_Update<'a>

Set value of MR2S field.

fn set_mr3i<'a>(&'a mut self, new_value: CT32B0_mcr_mr3i) -> CT32B0_mcr_Update<'a>

Set value of MR3I field.

fn set_mr3r<'a>(&'a mut self, new_value: CT32B0_mcr_mr3r) -> CT32B0_mcr_Update<'a>

Set value of MR3R field.

fn set_mr3s<'a>(&'a mut self, new_value: CT32B0_mcr_mr3s) -> CT32B0_mcr_Update<'a>

Set value of MR3S field.