Struct zinc_hal_lpc11xx::ioregs::CT32B0_tcr

pub struct CT32B0_tcr {
    // some fields omitted
}

Register TCR: Timer Control Register (TCR). The TCR is used to control the Timer Counter functions. The Timer Counter can be disabled or reset through the TCR.

Methods

impl CT32B0_tcr

fn get(&self) -> CT32B0_tcr_Get

Fetch the value of the TCR register.

fn cen(&self) -> bool

Get value of CEN field.

fn crst(&self) -> bool

Get value of CRST field.

fn ignoring_state(&self) -> CT32B0_tcr_Update

Create new updater that ignores current value of the TCR register.

fn set_cen<'a>(&'a mut self, new_value: bool) -> CT32B0_tcr_Update<'a>

Set value of CEN field.

fn set_crst<'a>(&'a mut self, new_value: bool) -> CT32B0_tcr_Update<'a>

Set value of CRST field.