Struct zinc_hal_lpc11xx::ioregs::C_CAN_canif1_arb2_Get

pub struct C_CAN_canif1_arb2_Get {
    // some fields omitted
}

CANIF1_ARB2: Message interface 1 arbitration 2

Methods

impl C_CAN_canif1_arb2_Get

fn new(reg: &C_CAN_canif1_arb2) -> C_CAN_canif1_arb2_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn id_28_16(&self) -> u32

Get value of ID_28_16 field: Message identifier 29-bit identifier (extended frame) 11-bit identifier (standard frame)

fn dir(&self) -> C_CAN_canif1_arb2_dir

Get value of DIR field: Message direction

fn xtd(&self) -> C_CAN_canif1_arb2_xtd

Get value of XTD field: Extend identifier

fn msgval(&self) -> C_CAN_canif1_arb2_msgval

Get value of MSGVAL field: Message valid The CPU must reset the MSGVAL bit of all unused Messages Objects during the initialization before it resets bit INIT in the CAN Control Register. This bit must also be reset before the identifier ID28:0, the control bits XTD, DIR, or the Data Length Code DLC3:0 are modified, or if the Messages Object is no longer required.