Struct zinc_hal_lpc11xx::ioregs::C_CAN_canif2_mctrl_Get

pub struct C_CAN_canif2_mctrl_Get {
    // some fields omitted
}

CANIF2_MCTRL: Message interface 1 message control

Methods

impl C_CAN_canif2_mctrl_Get

fn new(reg: &C_CAN_canif2_mctrl) -> C_CAN_canif2_mctrl_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn dlc_3_0(&self) -> u32

Get value of DLC_3_0 field: Data length code The Data Length Code of a Message Object must be defined the same as in all the corresponding objects with the same identifier at other nodes. When the Message Handler stores a data frame, it will write the DLC to the value given by the received message. 0000 - 1000 = Data frame has 0 - 8 data bytes. 1001 - 1111 = Data frame has 8 data bytes.

fn eob(&self) -> C_CAN_canif2_mctrl_eob

Get value of EOB field: End of buffer

fn txrqst(&self) -> C_CAN_canif2_mctrl_txrqst

Get value of TXRQST field: Transmit request

fn rmten(&self) -> C_CAN_canif2_mctrl_rmten

Get value of RMTEN field: Remote enable

fn rxie(&self) -> C_CAN_canif2_mctrl_rxie

Get value of RXIE field: Receive interrupt enable

fn txie(&self) -> C_CAN_canif2_mctrl_txie

Get value of TXIE field: Transmit interrupt enable

fn umask(&self) -> C_CAN_canif2_mctrl_umask

Get value of UMASK field: Use acceptance mask If UMASK is set to 1, the message object's mask bits have to be programmed during initialization of the message object before MAGVAL is set to 1.

fn intpnd(&self) -> C_CAN_canif2_mctrl_intpnd

Get value of INTPND field: Interrupt pending

fn msglst(&self) -> C_CAN_canif2_mctrl_msglst

Get value of MSGLST field: Message lost (only valid for message objects in the direction receive).

fn newdat(&self) -> C_CAN_canif2_mctrl_newdat

Get value of NEWDAT field: New data