Struct zinc_hal_lpc11xx::ioregs::C_CAN_canstat_Get

pub struct C_CAN_canstat_Get {
    // some fields omitted
}

CANSTAT: Status register

Methods

impl C_CAN_canstat_Get

fn new(reg: &C_CAN_canstat) -> C_CAN_canstat_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn lec(&self) -> C_CAN_canstat_lec

Get value of LEC field: Last error code Type of the last error to occur on the CAN bus.The LEC field holds a code which indicates the type of the last error to occur on the CAN bus. This field will be cleared to 0 when a message has been transferred (reception or transmission) without error. The unused code 111 may be written by the CPU to check for updates.

fn txok(&self) -> C_CAN_canstat_txok

Get value of TXOK field: Transmitted a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

fn rxok(&self) -> C_CAN_canstat_rxok

Get value of RXOK field: Received a message successfully This bit is reset by the CPU. It is never reset by the CAN controller.

fn epass(&self) -> C_CAN_canstat_epass

Get value of EPASS field: Error passive

fn ewarn(&self) -> C_CAN_canstat_ewarn

Get value of EWARN field: Warning status

fn boff(&self) -> C_CAN_canstat_boff

Get value of BOFF field: Busoff status