Struct zinc_hal_lpc11xx::ioregs::I2C_conset_Get

pub struct I2C_conset_Get {
    // some fields omitted
}

CONSET: I2C Control Set Register. When a one is written to a bit of this register, the corresponding bit in the I2C control register is set. Writing a zero has no effect on the corresponding bit in the I2C control register.

Methods

impl I2C_conset_Get

fn new(reg: &I2C_conset) -> I2C_conset_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn aa(&self) -> bool

Get value of AA field: Assert acknowledge flag.

fn si(&self) -> bool

Get value of SI field: I2C interrupt flag.

fn sto(&self) -> bool

Get value of STO field: STOP flag.

fn sta(&self) -> bool

Get value of STA field: START flag.

fn i2en(&self) -> bool

Get value of I2EN field: I2C interface enable.