Struct zinc_hal_lpc11xx::ioregs::SPI0_ris_Get

pub struct SPI0_ris_Get {
    // some fields omitted
}

RIS: Raw Interrupt Status Register

Methods

impl SPI0_ris_Get

fn new(reg: &SPI0_ris) -> SPI0_ris_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn rorris(&self) -> bool

Get value of RORRIS field: This bit is 1 if another frame was completely received while the RxFIFO was full. The ARM spec implies that the preceding frame data is overwritten by the new frame data when this occurs.

fn rtris(&self) -> bool

Get value of RTRIS field: This bit is 1 if the Rx FIFO is not empty, and has not been read for a time-out period. The time-out period is the same for master and slave modes and is determined by the SSP bit rate: 32 bits at PCLK / (CPSDVSR x [SCR+1]).

fn rxris(&self) -> bool

Get value of RXRIS field: This bit is 1 if the Rx FIFO is at least half full.

fn txris(&self) -> bool

Get value of TXRIS field: This bit is 1 if the Tx FIFO is at least half empty.