Struct zinc_hal_lpc11xx::ioregs::UART

pub struct UART;

Product name title=UM10398 Chapter title=LPC111x/LPC11Cxx UART Modification date=9/19/2011 Major revision=7 Minor revision=not available

Methods

impl UART

fn rbr() -> &'static mut UART_rbr

Receiver Buffer Register. Contains the next received character to be read. (DLAB=0)

fn thr() -> &'static mut UART_thr

Transmit Holding Register. The next character to be transmitted is written here. (DLAB=0)

fn dll() -> &'static mut UART_dll

Divisor Latch LSB. Least significant byte of the baud rate divisor value. The full divisor is used to generate a baud rate from the fractional rate divider. (DLAB=1)

fn dlm() -> &'static mut UART_dlm

Divisor Latch MSB. Most significant byte of the baud rate divisor value. The full divisor is used to generate a baud rate from the fractional rate divider. (DLAB=1)

fn ier() -> &'static mut UART_ier

Interrupt Enable Register. Contains individual interrupt enable bits for the 7 potential UART interrupts. (DLAB=0)

fn iir() -> &'static mut UART_iir

Interrupt ID Register. Identifies which interrupt(s) are pending.

fn fcr() -> &'static mut UART_fcr

FIFO Control Register. Controls UART FIFO usage and modes.

fn lcr() -> &'static mut UART_lcr

Line Control Register. Contains controls for frame formatting and break generation.

fn mcr() -> &'static mut UART_mcr

Modem control register

fn lsr() -> &'static mut UART_lsr

Line Status Register. Contains flags for transmit and receive status, including line errors.

fn msr() -> &'static mut UART_msr

Modem status register

fn scr() -> &'static mut UART_scr

Scratch Pad Register. Eight-bit temporary storage for software.

fn acr() -> &'static mut UART_acr

Auto-baud Control Register. Contains controls for the auto-baud feature.

fn fdr() -> &'static mut UART_fdr

Fractional Divider Register. Generates a clock input for the baud rate divider.

fn ter() -> &'static mut UART_ter

Transmit Enable Register. Turns off UART transmitter for use with software flow control.

fn rs485ctrl() -> &'static mut UART_rs485ctrl

RS-485/EIA-485 Control. Contains controls to configure various aspects of RS-485/EIA-485 modes.

fn rs485adrmatch() -> &'static mut UART_rs485adrmatch

RS-485/EIA-485 address match. Contains the address match value for RS-485/EIA-485 mode.

fn rs485dly() -> &'static mut UART_rs485dly

RS-485/EIA-485 direction control delay.