Struct zinc_hal_lpc11xx::ioregs::UART_ier_Get

pub struct UART_ier_Get {
    // some fields omitted
}

IER: Interrupt Enable Register. Contains individual interrupt enable bits for the 7 potential UART interrupts. (DLAB=0)

Methods

impl UART_ier_Get

fn new(reg: &UART_ier) -> UART_ier_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn rbrie(&self) -> UART_ier_rbrie

Get value of RBRIE field: RBR Interrupt Enable. Enables the Receive Data Available interrupt for UART. It also controls the Character Receive Time-out interrupt.

fn threie(&self) -> UART_ier_threie

Get value of THREIE field: THRE Interrupt Enable. Enables the THRE interrupt for UART. The status of this interrupt can be read from LSR[5].

fn rxlie(&self) -> UART_ier_rxlie

Get value of RXLIE field: RX Line Interrupt Enable. Enables the UART RX line status interrupts. The status of this interrupt can be read from LSR[4:1].

fn abeointen(&self) -> UART_ier_abeointen

Get value of ABEOINTEN field: Enables the end of auto-baud interrupt.

fn abtointen(&self) -> UART_ier_abtointen

Get value of ABTOINTEN field: Enables the auto-baud time-out interrupt.