Struct zinc_hal_lpc11xx::ioregs::UART_iir_Get

pub struct UART_iir_Get {
    // some fields omitted
}

IIR: Interrupt ID Register. Identifies which interrupt(s) are pending.

Methods

impl UART_iir_Get

fn new(reg: &UART_iir) -> UART_iir_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn intstatus(&self) -> UART_iir_intstatus

Get value of INTSTATUS field: Interrupt status. Note that IIR[0] is active low. The pending interrupt can be determined by evaluating IIR[3:1].

fn intid(&self) -> UART_iir_intid

Get value of INTID field: Interrupt identification. IER[3:1] identifies an interrupt corresponding to the UART Rx FIFO. All other combinations of IER[3:1] not listed below are reserved (100,101,111).

fn fifoenable(&self) -> u32

Get value of FIFOENABLE field: These bits are equivalent to FCR[0].

fn abeoint(&self) -> bool

Get value of ABEOINT field: End of auto-baud interrupt. True if auto-baud has finished successfully and interrupt is enabled.

fn abtoint(&self) -> bool

Get value of ABTOINT field: Auto-baud time-out interrupt. True if auto-baud has timed out and interrupt is enabled.