Struct zinc_hal_lpc11xx::ioregs::UART_mcr_Get

pub struct UART_mcr_Get {
    // some fields omitted
}

MCR: Modem control register

Methods

impl UART_mcr_Get

fn new(reg: &UART_mcr) -> UART_mcr_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn dtrc(&self) -> bool

Get value of DTRC field: DTR Control. Source for modem output pin, DTR. This bit reads as 0 when modem loopback mode is active.

fn rtsc(&self) -> bool

Get value of RTSC field: RTS Control. Source for modem output pin RTS. This bit reads as 0 when modem loopback mode is active.

fn lms(&self) -> bool

Get value of LMS field: Loopback Mode Select. The modem loopback mode provides a mechanism to perform diagnostic loopback testing. Serial data from the transmitter is connected internally to serial input of the receiver. Input pin, RXD, has no effect on loopback and output pin, TXD is held in marking state. The four modem inputs (CTS, DSR, RI and DCD) are disconnected externally. Externally, the modem outputs (RTS, DTR) are set inactive. Internally, the four modem outputs are connected to the four modem inputs. As a result of these connections, the upper four bits of the MSR will be driven by the lower four bits of the MCR rather than the four modem inputs in normal mode. This permits modem status interrupts to be generated in loopback mode by writing the lower four bits of MCR.

fn rtsen(&self) -> UART_mcr_rtsen

Get value of RTSEN field: RTS flow control

fn ctsen(&self) -> UART_mcr_ctsen

Get value of CTSEN field: CTS flow control