Struct zinc_hal_lpc11xx::ioregs::UART_msr_Get

pub struct UART_msr_Get {
    // some fields omitted
}

MSR: Modem status register

Methods

impl UART_msr_Get

fn new(reg: &UART_msr) -> UART_msr_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn dcts(&self) -> UART_msr_dcts

Get value of DCTS field: Delta CTS. Set upon state change of input CTS. Cleared on a MSR read.

fn ddsr(&self) -> UART_msr_ddsr

Get value of DDSR field: Delta DSR. Set upon state change of input DSR. Cleared on a MSR read.

fn teri(&self) -> UART_msr_teri

Get value of TERI field: Trailing Edge RI. Set upon low to high transition of input RI. Cleared on a MSR read.

fn ddcd(&self) -> UART_msr_ddcd

Get value of DDCD field: Delta DCD. Set upon state change of input DCD. Cleared on a MSR read.

fn cts(&self) -> bool

Get value of CTS field: Clear To Send State. Complement of input signal CTS. This bit is connected to MCR[1] in modem loopback mode.

fn dsr(&self) -> bool

Get value of DSR field: Data Set Ready State. Complement of input signal DSR. This bit is connected to MCR[0] in modem loopback mode.

fn ri(&self) -> bool

Get value of RI field: Ring Indicator State. Complement of input RI. This bit is connected to MCR[2] in modem loopback mode.

fn dcd(&self) -> bool

Get value of DCD field: Data Carrier Detect State. Complement of input DCD. This bit is connected to MCR[3] in modem loopback mode.