Struct zinc_hal_lpc11xx::ioregs::UART_rs485ctrl_Get

pub struct UART_rs485ctrl_Get {
    // some fields omitted
}

RS485CTRL: RS-485/EIA-485 Control. Contains controls to configure various aspects of RS-485/EIA-485 modes.

Methods

impl UART_rs485ctrl_Get

fn new(reg: &UART_rs485ctrl) -> UART_rs485ctrl_Get

Create a getter reflecting the current value of the given register.

fn raw(&self) -> u32

Get the raw value of the register.

fn nmmen(&self) -> UART_rs485ctrl_nmmen

Get value of NMMEN field: NMM enable.

fn rxdis(&self) -> UART_rs485ctrl_rxdis

Get value of RXDIS field: Receiver enable.

fn aaden(&self) -> UART_rs485ctrl_aaden

Get value of AADEN field: AAD enable.

fn sel(&self) -> UART_rs485ctrl_sel

Get value of SEL field: Select direction control pin

fn dctrl(&self) -> UART_rs485ctrl_dctrl

Get value of DCTRL field: Auto direction control enable.

fn oinv(&self) -> UART_rs485ctrl_oinv

Get value of OINV field: Polarity control. This bit reverses the polarity of the direction control signal on the RTS (or DTR) pin.