Struct zinc_hal_lpc11xx::ioregs::WWDT

pub struct WWDT;

Product name title=UM10398 Chapter title=LPC111x/LPC11Cxx Windowed WatchDog Timer (WDT) Modification date=9/19/2011 Major revision=6 Minor revision=not available

Methods

impl WWDT

fn wdmod() -> &'static mut WWDT_wdmod

Watchdog mode register. This register contains the basic mode and status of the Watchdog Timer.

fn wdtc() -> &'static mut WWDT_wdtc

Watchdog timer constant register. This register determines the time-out value.

fn wdfeed() -> &'static mut WWDT_wdfeed

Watchdog feed sequence register. Writing 0xAA followed by 0x55 to this register reloads the Watchdog timer with the value contained in WDTC.

fn wdtv() -> &'static mut WWDT_wdtv

Watchdog timer value register. This register reads out the current value of the Watchdog timer.

fn wdwarnint() -> &'static mut WWDT_wdwarnint

Watchdog Warning Interrupt compare value.

fn wdwindow() -> &'static mut WWDT_wdwindow

Watchdog Window compare value.