Struct zinc_hal_lpc11xx::ioregs::WWDT_wdmod

pub struct WWDT_wdmod {
    // some fields omitted
}

Register WDMOD: Watchdog mode register. This register contains the basic mode and status of the Watchdog Timer.

Methods

impl WWDT_wdmod

fn get(&self) -> WWDT_wdmod_Get

Fetch the value of the WDMOD register.

fn wden(&self) -> WWDT_wdmod_wden

Get value of WDEN field.

fn wdreset(&self) -> WWDT_wdmod_wdreset

Get value of WDRESET field.

fn wdtof(&self) -> bool

Get value of WDTOF field.

fn wdint(&self) -> bool

Get value of WDINT field.

fn wdprotect(&self) -> WWDT_wdmod_wdprotect

Get value of WDPROTECT field.

fn ignoring_state(&self) -> WWDT_wdmod_Update

Create new updater that ignores current value of the WDMOD register.

fn set_wden<'a>(&'a mut self, new_value: WWDT_wdmod_wden) -> WWDT_wdmod_Update<'a>

Set value of WDEN field.

fn set_wdreset<'a>(&'a mut self, new_value: WWDT_wdmod_wdreset) -> WWDT_wdmod_Update<'a>

Set value of WDRESET field.

fn set_wdtof<'a>(&'a mut self, new_value: bool) -> WWDT_wdmod_Update<'a>

Set value of WDTOF field.

fn set_wdint<'a>(&'a mut self, new_value: bool) -> WWDT_wdmod_Update<'a>

Set value of WDINT field.

fn set_wdprotect<'a>(&'a mut self, new_value: WWDT_wdmod_wdprotect) -> WWDT_wdmod_Update<'a>

Set value of WDPROTECT field.